requirements.txt
Django==1.8.1
MySQL-python==1.2.5
argparse==1.2.1
django-filter==0.9.2
django-grappelli==2.6.4
djangorestframework==3.1.1
wsgiref==0.1.2
